Added CommentCellViewModel
