+++ Kickstarter-iOS/Views/Cells/CommentPostFailedCell.swift
@@ -6,6 +6,8 @@ import UIKit
 final class CommentPostFailedCell: UITableViewCell, ValueCell {
   // MARK: - Properties
 
+  fileprivate let viewModel = CommentCellViewModel()
+
   private lazy var rootStackView = {
     UIStackView(frame: .zero)
       |> \.translatesAutoresizingMaskIntoConstraints .~ false
@@ -67,7 +69,7 @@ final class CommentPostFailedCell: UITableViewCell, ValueCell {
 
   internal func configureWith(value: DemoComment) {
     self.commentCellHeaderStackView.configureWith(comment: value)
-    self.commentLabel.text = value.body
+    self.viewModel.inputs.configureWith(comment: value)
   }
 
   private func configureViews() {
@@ -90,6 +92,12 @@ final class CommentPostFailedCell: UITableViewCell, ValueCell {
       self.separatorView.leadingAnchor.constraint(equalTo: self.contentView.leadingAnchor),
       self.separatorView.trailingAnchor.constraint(equalTo: self.contentView.trailingAnchor)
     ])
+  }
+
+  // MARK: - View model
+
+  internal override func bindViewModel() {
+    self.commentLabel.rac.text = self.viewModel.outputs.body
   }
 }
